% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForPackage09242019.R
\name{BH.Estimates}
\alias{BH.Estimates}
\title{Estimates for Breslow-Holubkov approach from the "osDesign" package}
\usage{
BH.Estimates(fmla = fmla, dat = dat, tps.fit = tps.fit)
}
\arguments{
\item{fmla}{The formula used in the "tps" call from "osDesign"}

\item{dat}{The phase 1 dataset coded with delta=1 for phase 2 sampled observations}

\item{tps.fit}{The fit output by the tps function in "osDesign"}
}
\value{
Estimates of marginalized risk difference, relative risk and odds ratios
}
\description{
Provides estimates using the BH approach for two stage designs for the marginal means, marginal risk difference, and marginal relative risk.  It requires as input the output object from the "tps" function from the "osDesign" package
}
\examples{
See package vignette
BH.Estimates
}
