% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForPackage09242019.R
\name{weighted.tmle.predicted.values}
\alias{weighted.tmle.predicted.values}
\title{IPCWTMLE2Phase

IPCW-TMLE to output the TMLEs of predicted values}
\usage{
weighted.tmle.predicted.values(w, a, y, delta, Q.SL.library, g.SL.library,
  wgts, max.wgt = Inf)
}
\arguments{
\item{w}{The covariates}

\item{a}{The treatment or exposure binary variable.}

\item{y}{The outcome}

\item{Q.SL.library}{SuperLearner library of algorithms for Q estimation}

\item{g.SL.library}{SuperLearner library of algorithms for g estimation}

\item{wgts}{observation level weights; should be 0 for delta=0}
}
\value{
fitted.values The predicted values for each observation from the TMLE
}
\description{
IPCW-TMLE outputting TMLEs of predicted values. This function only performs the TMLE step for targeting the predicted values. And does not output any additional estimates.
}
\examples{
weighted.tmle.predicted.values
}
