% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForPackage09242019.R
\name{weighted.tmle.separate}
\alias{weighted.tmle.separate}
\title{IPCWTMLE2Phase
Main IPCW-TMLE function for two-phase studies}
\usage{
weighted.tmle.separate(w, a, y, delta, Q.SL.library, g.SL.library, wgts,
  max.wgt = Inf)
}
\arguments{
\item{w}{The covariates}

\item{a}{The treatment or exposure binary variable.}

\item{y}{The outcome}

\item{delta}{Indicator for inclusion into the second phase}

\item{Q.SL.library}{SuperLearner library of algorithms for Q estimation}

\item{g.SL.library}{SuperLearner library of algorithms for g estimation}

\item{wgts}{Observation level weights; should be 0 for delta=0}
}
\value{
Estimates of marginalized risk difference, RR, and OR
}
\description{
IPCW-TMLE function to estimate E[E[Y|A=1]] and E[E[Y|A=0]] separately, and to also estimate the marginalized risk difference, relative risk, and odds ratio.  This function assumes that the full phase one data set is input with weights equal to 0 for those observations not selected for the second phase.  An indicator variable, delta, should equal 1 for all phase-two data and 0 for all other observations (non-phase-two).
}
\examples{
weighted.tmle.separate
}
